Move LocationMonitor to hooks dir and rename the file accordingly
