+++ code/core/src/manager/hooks/useLocation.ts
